Working version

+++ lab4/zad3/sender.c
@@ -3,10 +3,6 @@
 #include <stdlib.h>
 #include <unistd.h>
 #include <sys/types.h>
-#include <sys/wait.h>
-
-int sigqueue(pid_t pid, int sig, const union sigval value);
-
 
 void send_signal(pid_t pid, int signal_num, int value) {
     union sigval sigval_value;
@@ -14,67 +10,56 @@ void send_signal(pid_t pid, int signal_num, int value) {
     sigqueue(pid, signal_num, sigval_value);
 }
 
-void wait_for_signal(int signal_num) {
-    sigset_t sigset;
-    sigemptyset(&sigset);
-    sigaddset(&sigset, signal_num);
-    while(1) {
-        sigsuspend(&sigset);
-        return;
-    }
+void handler(int sig, siginfo_t* info, void* context)
+{
+    printf("S: Confirmation received.\n");
 }
 
-int main(int n, char *args[]) {
+int main(int n, char **args) {
      if(n < 2) {
-        printf("Usage: ./sender <catcher_pid>");
+        printf("Usage: ./sender <catcher_pid>\n");
         exit(1);
     }
 
     pid_t catcher_pid = atoi(args[1]);
-    /*int time_span = min(2, atoi(args[2]));
+
+    printf("------------- Sender started.--------------\n");
+    printf("S: Pid: %d\n", getpid());
+    printf("S: Catcher pid: %d\n", catcher_pid);
+    printf("S: Enter numbers in the range 1-5:\n");
 
     int modes[n-2];
-
     for(int i = 2; i < n; i++) {
         modes[i-2] = atoi(args[i]);
 
         if(modes[i-2] < 1 || modes[i-2]> 5)
         {
-            perror("Data Error. Ending program.");
+            perror("Data Error. Ending program.\n");
             send_signal(catcher_pid, SIGUSR1, 5);
-            wait_for_signal(SIGUSR1);
             exit(1);
         }
     }
 
+    struct sigaction act;
+    act.sa_flags = SA_SIGINFO;
+    act.sa_handler = handler;
+    if (sigaction(SIGUSR1, &act, NULL) < 0) {
+        perror("S: Sigaction error.\n");
+        exit(1);
+    }
+
     for(int i = 0; i < n-2; i++) {
         int mode = modes[i];
-        printf("Sending mode %d\n", mode);
+        printf("S: Sending mode %d\n", mode);
         send_signal(catcher_pid, SIGUSR1, mode);
-        wait_for_signal(SIGUSR1);
-    } */
-
-
-    int mode;
-
-    printf("Enter numbers in the range 1-5:\n");
-
-    while (scanf("%d", &mode) == 1) {
-        if (mode >= 1 && mode <= 5)
-        {
-            printf("Sending mode %d\n", mode);
-            send_signal(catcher_pid, SIGUSR1, mode);
-            wait_for_signal(SIGUSR1);
-        }
-        else 
-            printf("Invalid number. Please enter a number in the range 1-5.\n");
-
-        if(mode == 5)
-            break;
+        printf("S: Signal send %d\n", mode);
+        sigsuspend(&act.sa_mask);
+        printf("S: Signal received %d\n", mode);
     }
 
 
-    printf("All modes sent successfully!\n");
+    printf("S: All modes sent successfully!\n");
 
     return 0;
 }
+
